requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
Faker==13.3.1
drf_yasg==1.20.0
